Remove HidlInstrumentor logic entirely.
This userdebug-only feature is used for collecting profiling information
and getting coverage data. It was removed from user builds because it
isn't threadsafe, but also, it is causing a leak because it repeatedly
calls add_sysprop_change_callback.

For coverage information, we have no existing multi-process solution,
and we are focusing on in-process fuzzing.

For profiling information, perfetto/heapprofd/atrace/simpleperf and
others should be prefered.

HIDL is on the deprecation route, and unfortunately this feature has
bitrot, since the old VTS framework is deprecated, and we use a unified
tradefed now. If anyone needs this functionality still, please reach out
to me or file an issue. Thanks!

Note - we can't remove the HidlInstrumentor class itself due to
its ABI being frozen.

Fixes: 202852241
Test: build only
Change-Id: Ibce1f5762740570e6d9405a49c4d917ffb078152